feat: Removed obsolete MJSONWP touch actions
