+++ lib/commands/actions.js
@@ -43,5 +43,44 @@ export async function mobileUnscheduleAction(opts) {
 }
 
 /**
+ * @this {AndroidUiautomator2Driver}
+ * @param {import('@appium/types').StringRecord[]} actions
+ * @returns {Promise<void>}
+ */
+export async function performActions(actions) {
+  // This is mandatory, since Selenium API uses MOUSE as the default pointer type
+  const preprocessedActions = actions.map((action) =>
+    Object.assign(
+      {},
+      action,
+      action.type === 'pointer'
+        ? {
+            parameters: {
+              pointerType: 'touch',
+            },
+          }
+        : {}
+    )
+  );
+  this.log.debug(`Preprocessed actions: ${JSON.stringify(preprocessedActions, null, '  ')}`);
+  await this.uiautomator2.jwproxy.command(
+    '/actions',
+    'POST',
+    {
+      actions: preprocessedActions,
+    }
+  );
+}
+
+/**
+ * @this {AndroidUiautomator2Driver}
+ * @returns {Promise<void>}
+ */
+// eslint-disable-next-line require-await
+export async function releaseActions() {
+  this.log.info('On this platform, releaseActions is a no-op');
+}
+
+/**
  * @typedef {import('../driver').AndroidUiautomator2Driver} AndroidUiautomator2Driver
  */
