+++ lib/commands/gestures.js
@@ -250,7 +250,7 @@ export async function mobileScroll(opts) {
     );
   }
   await this.uiautomator2.jwproxy.command(
-    '/touch/scroll',
+    '/gestures/scroll_to',
     'POST',
     {
       origin: toOrigin(elementId || element),
