+++ package.json
@@ -59,7 +59,7 @@
     "appium-adb": "^12.0.0",
     "appium-android-driver": "^8.4.0",
     "appium-chromedriver": "^5.6.28",
-    "appium-uiautomator2-server": "^6.0.7",
+    "appium-uiautomator2-server": "^7.0.0",
     "asyncbox": "^3.0.0",
     "axios": "^1.6.5",
     "bluebird": "^3.5.1",
